requirements.txt
flask
requests
PyMuPDF
Pillow 